:sparkles: refactoring: presentation 패키지 구조 Domain 별로 변경
