+++ app/src/main/java/com/pickpick/pickpick/presentation/auth/findemail/viewmodel/FindEmailUiState.kt
@@ -1,0 +1,8 @@
+package com.pickpick.pickpick.presentation.auth.findemail.viewmodel
+
+import com.pickpick.pickpick.core.presentation.UiState
+
+data class FindEmailUiState(
+    val email: String = "",
+    val errorText: String? = null
+) : UiState
