+++ app/src/main/java/com/pickpick/pickpick/presentation/auth/findpassword/viewmodel/FindPasswordUiState.kt
@@ -1,0 +1,8 @@
+package com.pickpick.pickpick.presentation.auth.findpassword.viewmodel
+
+import com.pickpick.pickpick.core.presentation.UiState
+
+data class FindPasswordUiState(
+    val email: String = "",
+    val errorText: String? = null
+) : UiState
