+++ app/src/main/java/com/pickpick/pickpick/presentation/auth/login/viewmodel/LoginUiState.kt
@@ -1,0 +1,10 @@
+package com.pickpick.pickpick.presentation.auth.login.viewmodel
+
+import com.pickpick.pickpick.core.presentation.UiState
+
+data class LoginUiState(
+    val email: String = "",
+    val password: String = "",
+    val errorText: String? = null,
+    val enabled: Boolean = false,
+) : UiState
