--- app/src/main/java/com/pickpick/pickpick/presentation/findpassword/viewmodel/FindPasswordUiState.kt
@@ -1,8 +1,0 @@
-package com.pickpick.pickpick.presentation.findpassword.viewmodel
-
-import com.pickpick.pickpick.core.presentation.UiState
-
-data class FindPasswordUiState(
-    val email: String = "",
-    val errorText: String? = null
-) : UiState
